feat: added context and the opt in
